[containerisation]

+++ backend/controller.py
@@ -1,17 +1,18 @@
 from __future__ import annotations
 import asyncio
-import os, dotenv
+import os
+from dotenv import load_dotenv
+from pathlib import Path    
 from typing import Awaitable, Callable
 
 
-
-dotenv.load_dotenv("backend/.env")
+BACKENDDIR = Path(__file__).resolve().parent
+load_dotenv(BACKENDDIR.parent / ".env")
 
 from contextlib import asynccontextmanager
 from fastapi import FastAPI, Request, responses, staticfiles
 from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy import text
-from pathlib import Path
 
 from .db import *
 from .logger import logger
@@ -23,14 +24,16 @@ from .players import players_router
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     async with db_engine.begin() as conn:
-
-        await conn.execute(text("DROP SCHEMA public CASCADE"))
-        await conn.execute(text("CREATE SCHEMA public"))
+        
+        if os.getenv("DROP_DB_ON_STARTUP", "False").lower() in ("true", "1", "yes"):
+            logger.warning("Dropping and recreating public schema on startup as per DROP_DB_ON_STARTUP env var")
+            await conn.execute(text("DROP SCHEMA public CASCADE"))
+            await conn.execute(text("CREATE SCHEMA public"))
 
         await conn.run_sync(Base.metadata.create_all)
 
-        # Set eager task factory for websockets
 
+    # Set eager task factory for websockets
     asyncio.get_running_loop().set_task_factory(asyncio.eager_task_factory)
  
     asyncio.create_task(cleaner.run_passive_cleanup()) # start the passive cleaner
@@ -45,19 +48,22 @@ app = FastAPI(lifespan=lifespan)
 app.include_router(games_router)
 app.include_router(players_router)
 
-allowed = os.getenv("CORS_ALLOW_ORIGINS", "*").split(",")
 
-logger.info(f"CORS allowed origins: {allowed}")
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=allowed,
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
+if allowed := os.getenv("CORS_ALLOW_ORIGINS"):
+    allowed = [origin.strip() for origin in allowed.split(",")]
 
-BACKENDDIR = Path(__file__).parent
+    logger.info(f"CORS allowed origins: {allowed}")
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=allowed,
+        allow_credentials=False,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+
+
 FROTENTDIR = BACKENDDIR.parent / "frontend/dist"
+
 app.mount(
     "/assets", staticfiles.StaticFiles(directory=FROTENTDIR / "assets"), name="assets"
 )
@@ -66,6 +72,7 @@ app.mount(
 
 @app.middleware("http")
 async def noise_filter(request: Request, call_next: Callable[[Request], Awaitable[responses.Response]]):
+    """Catching noisy requests"""
     path = request.url.path
     if ".well-known" in path or path.endswith(".php"):
         return responses.Response(status_code=404)
